requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
cryptography==41.0.2
DateTime==5.2
dnspython==2.4.0
email-validator==2.0.0.post2
fastapi==0.100.0
fastapi-users-db-sqlalchemy==6.0.0
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
idna==3.4
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
passlib==1.7.4
pycparser==2.21
pydantic==2.0.2
pydantic_core==2.1.2
PyJWT==2.7.0
PySocks==1.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
spotipy==2.23.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
zope.interface==6.0
